Sun, Jan 16, 2022, 6:20 PM +00:00

+++ credit/credit.c
@@ -2,13 +2,13 @@
 #include <stdio.h>
 
 // Initialize my functions
-int alternative_digits(long int n);
+int alternative_digits(float n);
 
 
 
 int main(void)
 {
-	long int n = get_long("Numbers: ");
+	float n = get_long("Numbers: ");
 
 	int z = (n % 10) + alternative_digits(n);
 
@@ -17,9 +17,9 @@ int main(void)
 }
 
 
-int alternative_digits(long int n)
+int alternative_digits(float n)
 {
-	int y;
+	float y;
 	for(y = 3; (n/10^y)>=(0.1); y = y + 2)
 	{
 		int j = (n % 10^(y)/10^(y - 1)) - (n % 10^(y - 1))/(10^(y - 1));
